add nvme support to formatting tools

+++ bin/tdh-remote-format.sh
@@ -12,20 +12,22 @@ fi
 
 # -----------------------------------
 
-devtypes=( "sd" "xvd" )
+devtypes=( "sd" "nvme" )
 devtype="${devtypes[0]}"
 format="$TDH_FORMAT"
 
-pathpfx=
+pathpfx="data"
 volnum=1
 usexfs=0
 dryrun=0
 usegcp=0
+dtype=0
 ident=
 force=
 zone=
 user="$USER"
 
+# -----------------------------------
 
 usage()
 {
@@ -37,25 +39,32 @@ usage()
     echo "  -f|--force            : Set force option on 'mkfs'"
     echo "  -G|--use-gcp          : Use Google API for connecting to hosts"
     echo "  -h|--help             : Show usage info and exit"
-    echo "  -i|--indentity <file> : SSH Identity file to use for hosts"
+    echo "  -i|--identity <file>  : SSH Identity file."
     echo "  -n|--dryrun           : Enable dryrun (no actions are taken)"
     echo "  -p|--prefix  <path>   : Pathname prefix (default is /data)"
+    echo "  -t|--type    <0|1>    : Disk type, 0=Standard, 1=NVME"
     echo "  -u|--user    <name>   : Name of remote user if not '$user'"
     echo "  -x|--use-xfs          : Use XFS instead of default EXT4"
-    echo "  -z|--zone   <zoneid>  : GCP zone of target, if GCP and not default"
+    echo "  -z|--zone   <zoneid>  : GCP zone of target, if applicable."
     echo "  -V|--version          : Show version info and exit"
     echo ""
     echo "  eg. $TDH_PNAME -n 5 -x host1"
     echo "  Will format and mount 5 drives (sdb through sdf) "
-    echo "  as /data01 through /data05"
+    echo "  as /data01 through /data05 using XFS"
     echo ""
 }
 
+
+# -----------------------------------
+# MAIN
+
 rt=0
 chars=( {b..z} )
+nums=( {1..25} )
 maxvols=${#chars[@]}
 ssh="ssh"
 scp="scp"
+
 
 while [ $# -gt 0 ]; do
     case "$1" in
@@ -84,6 +93,10 @@ while [ $# -gt 0 ]; do
             pathpfx="$2"
             shift
             ;;
+        -t|--type)
+            dtype=$2
+            shift
+            ;;
         -u|--user)
             user="$2"
             shift
@@ -107,8 +120,23 @@ while [ $# -gt 0 ]; do
     shift
 done
 
+# -----------------------------------
+
 if [ -z "$hosts" ]; then
     usage
+    exit 1
+fi
+
+if [ $dtype -gt 1 ]; then
+    echo "Error in DiskType.."
+    exit 1
+fi 
+if [ $dtype -eq 1 ]; then
+    devtype="${devtypes[$dtype]}"
+fi
+
+if [ $volnum -gt $maxvols ]; then
+    echo "Error, --disknum '$volnum' greater than maximum allowed '$maxvols'"
     exit 1
 fi
 
@@ -122,8 +150,11 @@ if [ $usegcp -eq 1 ]; then
 else
     if [ -n "$ident" ]; then
         ssh="$ssh -i $ident"
+        scp="$scp -i $ident"
     fi
 fi
+
+# -----------------------------------
 
 hostssh=
 
@@ -144,11 +175,15 @@ for host in $hosts; do
     fi
 
     for (( i=0; i<$volnum; )); do
-        device="/dev/${devtype}${chars[i++]}"
+        if [ $dtype -eq 0 ]; then
+            device="/dev/${devtype}${chars[i++]}"
+        else
+            device="/dev/${devtype}${nums[i++]}n1"
+        fi
         dnum=$( printf "%02d" $i )
         mnt="/${pathpfx}${dnum}"
 
-        cmd="$hostssh './$format"
+        cmd="./$format"
 
         if [ $force -eq 1 ]; then
             cmd="$cmd --force"
@@ -158,11 +193,11 @@ for host in $hosts; do
             cmd="$cmd --use-xfs"
         fi
 
-        cmd="$cmd $device $mnt'"
+        cmd="$cmd $device $mnt"
 
-        echo "( $cmd )"
+        echo "( $hostssh \"$cmd\" )"
         if [ $dryrun -eq 0 ]; then
-            ( $cmd )
+            ( $hostssh "$cmd" )
             rt=$?
         fi
 
@@ -176,7 +211,9 @@ for host in $hosts; do
         break
     fi
 
-    ( $hostssh 'rm $format' )
+    if [ $dryrun -eq 0 ]; then
+        ( $hostssh "rm $format" )
+    fi
 done
 
 echo "$TDH_PNAME Finished."
